BusinessDomains/users-auth/auth-service/requirements.txt
Flask
Flask-SQLAlchemy
bcrypt
PyJWT